fixed from PVS-Studio:
V575 The number of processed elements should be passed to the 'GetWindowTextW' function as the third argument instead of buffer's size in bytes. urlctrl.cpp 166, 214
V611 The memory was allocated using 'new T[]' operator but was released using the 'delete' operator. Consider inspecting this code. It's probably better to use 'delete [] pBuffer;'. plugindefinition.cpp 133, 154